Trace state in lower part of window, not with presets ScrollViewer for main tab
