+++ ETWController/App.config
@@ -82,6 +82,12 @@
             <setting name="TraceOpenCmdLine" serializeAs="String">
                 <value>wpa -i $FileName -profile ETW\Overview.wpaProfile</value>
             </setting>
+            <setting name="CaptureMouseWheel" serializeAs="String">
+                <value>False</value>
+            </setting>
+            <setting name="CaptureMouseMove" serializeAs="String">
+                <value>False</value>
+            </setting>
         </ETWController.Configuration>
     </userSettings>
     <applicationSettings>
